Added buttons to FilteredBegRandom
